[explorer/rest] refactor: move PostgresTestDatabase tests to test_postgres_utils.py
