+++ explorer/rest/tests/test_symbol_core_routes.py
@@ -11,26 +11,9 @@ from rest import create_app, setup_symbol_facade
 from rest.facade.SymbolRestFacade import SymbolRestFacade
 from rest.model.common import DatabaseConfig
 
-from .test.EnvTestUtils import rest_settings_env, temporary_env_values
+from .test.EnvTestUtils import rest_settings_env
 from .test.SymbolBlockTestUtils import create_symbol_block, create_symbol_importance_block, create_symbol_sync_state
 from .test.SymbolHealthTestUtils import create_symbol_health
-
-
-class FakePostgresql:
-	def __init__(self):
-		self.stopped = False
-
-	@staticmethod
-	def dsn():
-		return {
-			'database': 'generated',
-			'user': 'postgres',
-			'host': '127.0.0.1',
-			'port': '5432'
-		}
-
-	def stop(self):
-		self.stopped = True
 
 
 def _create_config_file(
@@ -424,48 +407,3 @@ def test_symbol_facade_node_error(symbol_database_config):
 
 	# Assert:
 	assert str(exception_info.value) == 'Configured Symbol node host is not in SYMBOL_NODE_ALLOWED_HOSTS'
-
-
-def test_external_postgres_config():
-	# Arrange:
-	postgres_env = {
-		'EXPLORER_TEST_POSTGRES_HOST': 'postgres.example',
-		'EXPLORER_TEST_POSTGRES_DATABASE': 'symbol_test',
-		'EXPLORER_TEST_POSTGRES_USER': 'symbol_user',
-		'EXPLORER_TEST_POSTGRES_PORT': '15432',
-		'EXPLORER_TEST_POSTGRES_CREATE_DATABASE': 'false'
-	}
-
-	# Act:
-	with temporary_env_values(postgres_env):
-		with PostgresTestDatabase() as db_config:
-			result = db_config
-
-	# Assert:
-	assert DatabaseConfig(
-		'symbol_test',
-		'symbol_user',
-		'',
-		'postgres.example',
-		'15432') == result
-
-
-def test_testing_postgresql_fallback():
-	# Arrange:
-	fake_postgresql = FakePostgresql()
-
-	# Act:
-	with temporary_env_values({'EXPLORER_TEST_POSTGRES_HOST': None}):
-		with PostgresTestDatabase(
-			postgresql_factory=lambda: fake_postgresql
-		) as db_config:
-			result = db_config
-
-	# Assert:
-	assert DatabaseConfig(
-		'generated',
-		'postgres',
-		'',
-		'127.0.0.1',
-		'5432') == result
-	assert fake_postgresql.stopped
